changed MAX_PLAYERS to PLAYER_NUM and fixed all headers after moving constants.h and GameState.h
