+++ client/Window.h
@@ -5,7 +5,7 @@
 #include "shader.h"
 #include "Character.h"
 #include "EnvElement.h"
-#include "../common/constants.h"
+#include "../common/networking/CommunicationConstants.h"
 
 class Window
 {
